+++ Sources/StreamChatSwiftUI/ChatChannel/ChatChannelView.swift
@@ -73,6 +73,10 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
                             },
                             onJumpToMessage: viewModel.jumpToMessage(messageId:)
                         )
+                        .environment(\.highlightedMessageId, viewModel.highlightedMessageId)
+                        .dismissKeyboardOnTap(enabled: true) {
+                            hideComposerCommandsAndAttachmentsPicker()
+                        }
                         .overlay(
                             viewModel.currentDateString != nil ?
                                 factory.makeDateIndicatorView(dateString: viewModel.currentDateString!)
@@ -81,7 +85,9 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
                     } else {
                         ZStack {
                             factory.makeEmptyMessagesView(for: channel, colors: colors)
-                                .dismissKeyboardOnTap(enabled: keyboardShown)
+                                .dismissKeyboardOnTap(enabled: keyboardShown) {
+                                    hideComposerCommandsAndAttachmentsPicker()
+                                }
                             if viewModel.shouldShowTypingIndicator {
                                 factory.makeTypingIndicatorBottomView(
                                     channel: channel,
@@ -183,13 +189,6 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
             viewModel.reactionsShown = false
             messageDisplayInfo = nil
         }
-        .onChange(of: presentationMode.wrappedValue, perform: { newValue in
-            if newValue.isPresented == false {
-                viewModel.onViewDissappear()
-            } else {
-                viewModel.setActive()
-            }
-        })
         .background(
             Color(colors.background).background(
                 TabBarAccessor { _ in
@@ -220,10 +219,13 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
         let bottomPadding = topVC()?.view.safeAreaInsets.bottom ?? 0
         return bottomPadding
     }
-}
 
-extension PresentationMode: Equatable {
-    public static func == (lhs: PresentationMode, rhs: PresentationMode) -> Bool {
-        lhs.isPresented == rhs.isPresented
+    private func hideComposerCommandsAndAttachmentsPicker() {
+        NotificationCenter.default.post(
+            name: .attachmentPickerHiddenNotification, object: nil
+        )
+        NotificationCenter.default.post(
+            name: .commandsOverlayHiddenNotification, object: nil
+        )
     }
 }